+++ app/src/main/java/cl/agnov/ameli/ui/viewmodel/ViewModelFactories.kt
@@ -1,0 +1,22 @@
+package cl.agnov.ameli.ui.viewmodel
+
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.viewmodel.initializer
+import androidx.lifecycle.viewmodel.viewModelFactory
+import cl.agnov.ameli.AmeliApplication
+import cl.agnov.ameli.sip.LinphoneManager
+
+object ViewModelFactories {
+
+    val settings = viewModelFactory {
+        initializer {
+            val app = this[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY] as AmeliApplication
+            SettingsViewModel(
+                preferencesRepository = app.container.preferencesRepository,
+                secureCredentialStore = app.container.secureCredentialStore,
+                sipAccountManager = app.container.sipAccountManager,
+                registrationState = LinphoneManager.registrationState,
+            )
+        }
+    }
+}